add:stimulus transition
