+++ rails/app/javascript/controllers/index.js
@@ -24,3 +24,6 @@ application.register("modal", ModalController)
 
 import PopupController from "./popup_controller"
 application.register("popup", PopupController)
+
+import TransitionController from "./transition_controller"
+application.register("transition", TransitionController)
