+++ rails/app/views/beans/index.html.slim
@@ -16,7 +16,7 @@
           = f.submit class: "ui icon teal button",id: 'search' do
 
   - if @beans.present?
-    .ui.three.stackable.cards.segment.scroll#sidebar
+    .ui.three.stackable.cards.segment.scroll#sidebar data-controller= 'transition'
       .ui.right.close.rail
         .ui.sticky
           .ui.labeled.icon.teal.button#top-button
