+++ rails/app/views/histories/index.html.slim
@@ -8,7 +8,7 @@
       h1#title 閲覧履歴
 
   - if @histories.present?
-    .ui.segment.scroll#sidebar
+    .ui.segment.scroll#sidebar data-controller='transition'
       .ui.right.close.rail
         .ui.sticky
           .ui.labeled.icon.teal.button#top-button
