+++ rails/app/views/likes/index.html.slim
@@ -9,7 +9,7 @@
       p
         |全アイテムのお気に入り数をランキング順に表示しています。
 
-  .ui.three.stackable.cards.segment.mt-3rem data-controller='popup' data-action='connect->popup#ratepopup'
+  .ui.three.stackable.cards.segment.mt-3rem data-controller='popup transition' data-action='connect->popup#ratepopup'
     - @rankings.each.with_index(1) do |rank, i|
       a.ui.fluid.link.card.post.scale.rank-card href= product_path(rank.id)
         .ui.image
